Перенос всех view из shop.nordihome. в shop. Удален shop.nbrussia.

+++ database/migrations/2025_11_02_143830_update_contacts_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('contacts', function (Blueprint $table) {
-            $table->string('slug')->nullable();
-        });
+        if (!Schema::hasColumn('contacts', 'slug')) {
+            Schema::table('contacts', function (Blueprint $table) {
+                $table->string('slug')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('contacts', function (Blueprint $table) {
-            $table->dropColumn('slug');
-        });
+        if (Schema::hasColumn('contacts', 'slug')) {
+            Schema::table('contacts', function (Blueprint $table) {
+                $table->dropColumn('slug');
+            });
+        }
     }
 };
